Move entity stuff to modelling package instead of command package
