+++ integrationtests/src/test/java/org/axonframework/integrationtests/testsuite/student/commands/AssignMentorCommand.java
@@ -17,7 +17,7 @@
 package org.axonframework.integrationtests.testsuite.student.commands;
 
 import org.axonframework.integrationtests.testsuite.student.common.StudentMentorModelIdentifier;
-import org.axonframework.modelling.command.annotation.TargetEntityId;
+import org.axonframework.modelling.annotation.TargetEntityId;
 
 public record AssignMentorCommand(
         String menteeId,
