+++ modelling/src/main/java/org/axonframework/modelling/annotation/TargetEntityIdMemberMismatchException.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.axonframework.modelling.command.annotation;
+package org.axonframework.modelling.annotation;
 
 import jakarta.annotation.Nonnull;
